backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
PyJWT==2.8.0
pydantic==2.5.0
redis==5.0.1
PyYAML==6.0.1
pytest==7.4.3
pytest-flask==1.2.0